requirements.txt
python-twitter
colorama
grequests
Pillow
